requirements.txt
Flask
easyocr
numpy
opencv-python-headless
matplotlib
pillow